+++ src/main/java/lapesd/saturnus/model/dataStructures/CircularList.java
@@ -5,8 +5,8 @@ import java.util.ArrayList;
 public class CircularList<T> extends ArrayList<T> {
     private static int lastAccess = -1;
 
-    public T get(int index) {
-        return super.get(index % size());
+    public T get(long index) {
+        return super.get((int)index % size());
     }
 
     public T next() {
